+++ src/parser/effects/specialFeats.js
@@ -1,10 +1,12 @@
 import { baseItemEffect } from "./effects.js";
+import utils from "../../utils.js";
 
 /**
  * This function is mainly for effects that can't be dynamically generated
  * @param {*} document
  */
 export function featureEffectAdjustment(document) {
+  const midiQolInstalled = utils.isModuleInstalledAndActive("midi-qol");
   switch (document.name) {
     // if using active auras add the aura effect
     case "Aura of Courage":
@@ -25,12 +27,14 @@ export function featureEffectAdjustment(document) {
     case "Unarmored Movement": {
       document.effects.forEach((effect) => {
         if (effect.label.includes("Constant Effects")) {
-          effect.changes = [{
-            key: "data.attributes.movement.walk",
-            value: "max(10+(ceil(((@classes.monk.levels)-5)/4))*5,10)",
-            mode: CONST.ACTIVE_EFFECT_MODES.ADD,
-            priority: 20,
-          }];
+          effect.changes = [
+            {
+              key: "data.attributes.movement.walk",
+              value: "max(10+(ceil(((@classes.monk.levels)-5)/4))*5,10)",
+              mode: CONST.ACTIVE_EFFECT_MODES.ADD,
+              priority: 20,
+            },
+          ];
         }
       });
       break;
@@ -92,6 +96,96 @@ export function featureEffectAdjustment(document) {
       effect.flags.dae.transfer = false;
       effect.flags.dae.stackable = false;
       document.effects.push(effect);
+      break;
+    }
+    case "Potent Cantrip": {
+      let effect = baseItemEffect(document, `${document.name}`);
+      effect.changes.push({
+        key: "flags.midi-qol.potentCantrip",
+        value: "1",
+        mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM,
+        priority: 20,
+      });
+      document.effects.push(effect);
+      break;
+    }
+    case "Bardic Inspiration": {
+      if (midiQolInstalled) {
+        document.data.damage.parts = [];
+        let inspiredEffect = baseItemEffect(document, "Inspired");
+        inspiredEffect.changes.push(
+          {
+            key: "flags.midi-qol.optional.bardicInspiration.attack",
+            mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+            value: "1d@flags.dae.BardicInspirationDice",
+            priority: "20",
+          },
+          {
+            key: "flags.midi-qol.optional.bardicInspiration.save",
+            mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+            value: "1d@flags.dae.BardicInspirationDice",
+            priority: "20",
+          },
+          {
+            key: "flags.midi-qol.optional.bardicInspiration.check",
+            mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+            value: "1d@flags.dae.BardicInspirationDice",
+            priority: "20",
+          },
+          {
+            key: "flags.midi-qol.optional.bardicInspiration.label",
+            mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+            value: "Bardic Inspiration",
+            priority: "20",
+          }
+        );
+        inspiredEffect.transfer = false;
+        inspiredEffect.disabled = false;
+        setProperty(inspiredEffect, "flags.dae.transfer", false);
+        setProperty(inspiredEffect, "flags.dae.stackable", false);
+        setProperty(inspiredEffect, "flags.dae.macroRepeat", "none");
+        setProperty(inspiredEffect, "flags.dae.specialDuration", []);
+
+        if (document.flags.ddbimporter.subclass === "College of Valor") {
+          inspiredEffect.changes.push(
+            {
+              key: "flags.midi-qol.optional.bardicInspiration.damage",
+              mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+              value: "1d@flags.dae.BardicInspirationDice",
+              priority: "20",
+            },
+            {
+              key: "flags.midi-qol.optional.bardicInspiration.ac",
+              mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+              value: "1d@flags.dae.BardicInspirationDice",
+              priority: "20",
+            }
+          );
+        }
+
+        document.effects.push(inspiredEffect);
+
+        let diceEffect = baseItemEffect(document, "Bardic Inspiration Dice");
+        diceEffect.changes.push({
+          key: "flags.dae",
+          mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM,
+          value: "BardicInspirationDice (floor(@classes.bard.levels/5)+3) * 2",
+          priority: "20",
+        });
+        diceEffect.transfer = true;
+        diceEffect.disabled = false;
+        setProperty(diceEffect, "flags.dae.transfer", true);
+        setProperty(diceEffect, "flags.dae.stackable", false);
+        setProperty(diceEffect, "flags.dae.macroRepeat", "none");
+        setProperty(diceEffect, "flags.dae.specialDuration", []);
+        document.effects.push(diceEffect);
+        document.flags["midi-qol"] = {
+          onUseMacroName: "",
+          effectActivation: false,
+          forceCEOff: false,
+        };
+        console.warn(document);
+      }
       break;
     }
     // no default